fix conditions. prettified with artistic touch
